fix: use AccountOrganizationFiels fragment to prevent missing props
We do some manual cache changes when an organization is switched in
order to prevent over-fetching but if the fields in both
currentOrganization and organizations are not the same the query
write will fail to validate and break.

Using a fragment ensures that we request the same fields on both edges.